refactor: modularize schema and file discovery
